requirements.txt
azure-functions
requests
pypdf
python-docx